@@ -1,53 +1,44 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
-	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
-		<UserSecretsId>5c4af1e4-becd-4691-bad5-54e363aab345</UserSecretsId>
-		<TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
-	</PropertyGroup>
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <UserSecretsId>5c4af1e4-becd-4691-bad5-54e363aab345</UserSecretsId>
+        <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
+        <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
+        <DockerComposeProjectPath>..\docker-compose.dcproj</DockerComposeProjectPath>
+    </PropertyGroup>
 
-	<ItemGroup>
-		<Compile Remove="NewFolder\**" />
-		<Content Remove="NewFolder\**" />
-		<EmbeddedResource Remove="NewFolder\**" />
-		<None Remove="NewFolder\**" />
-	</ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="AspNet.Security.OAuth.Discord" Version="5.0.3" />
+        
+        <PackageReference Include="AutoMapper" Version="13.0.1" />
+        <!--<PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="8.1.1" />-->
+        
+        <PackageReference Include="Google.Apis.YouTube.v3" Version="1.52.0.2378" />
+        <PackageReference Include="Microsoft.AspNetCore.Authentication" Version="2.2.0" />
+        <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="5.0.9" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="5.0.4" />
+        
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.1">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>        
+        <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.1" />        
+        <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="8.0.1">
+          <PrivateAssets>all</PrivateAssets>
+          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        
+        <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.19.5" />
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.5.1" />
+        <PackageReference Include="MassTransit" Version="8.1.4-develop.1646" />
+        <PackageReference Include="MassTransit.SqlTransport.SqlServer" Version="8.1.4-develop.1646" />
+    </ItemGroup>
 
-	<ItemGroup>
-	  <None Remove="ClientApp\src\app\file.ts" />
-	  <None Remove="ClientApp\src\app\_enums\trackListType.ts" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="AspNet.Security.OAuth.Discord" Version="5.0.3" />
-		<PackageReference Include="AutoMapper" Version="10.1.1" />
-		<PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="8.1.1" />
-		<PackageReference Include="Discord.Net" Version="3.7.1" />
-		<PackageReference Include="Google.Apis.YouTube.v3" Version="1.52.0.2378" />
-		<PackageReference Include="Microsoft.AspNetCore.Authentication" Version="2.2.0" />
-		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="5.0.9" />
-		<PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="5.0.4" />
-		<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="6.0.5">
-			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-			<PrivateAssets>all</PrivateAssets>
-		</PackageReference>
-		<PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="6.0.5" />
-		<PackageReference Include="Swashbuckle.AspNetCore" Version="5.5.1" />
-		<PackageReference Include="Victoria" Version="5.2.8" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<Folder Include="Data\Migrations\" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\DisJockey.Core\DisJockey.Core.csproj" />
-		<ProjectReference Include="..\DisJockey.Discord\DisJockey.Discord.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <TypeScriptCompile Include="ClientApp\src\app\_enums\paginationType.ts" />
-	  <TypeScriptCompile Include="ClientApp\src\app\_enums\trackListType.ts" />
-	</ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\DisJockey.Core\DisJockey.Core.csproj" />
+        <ProjectReference Include="..\DisJockey.Services\DisJockey.Services.csproj" />
+        <!--<ProjectReference Include="..\DisJockey.Discord\DisJockey.Discord.csproj" />-->
+    </ItemGroup>
 
 </Project>
