@@ -7,37 +7,44 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace DisJockey {
-    public class Program {
-        public static async Task Main(string[] args) {
-            var host = CreateHostBuilder(args).Build();
+namespace DisJockey;
 
-            using (var scope = host.Services.CreateScope()) {
-                var services = scope.ServiceProvider;
+public class Program
+{
+    public static async Task Main(string[] args)
+    {
+        var host = CreateHostBuilder(args).Build();
 
-                try  {
-                    var context = services.GetRequiredService<DataContext>();
+        using (var scope = host.Services.CreateScope())
+        {
+            var services = scope.ServiceProvider;
 
-                    await context.Database.MigrateAsync();
+            try
+            {
+                var context = services.GetRequiredService<DataContext>();
 
-                    var env = services.GetRequiredService<IHostEnvironment>();
-                    if (env.IsDevelopment())
-                    {
-                        await Seed.SeedData(context);
-                    }
-                } catch (Exception e) {
-                    var logger = services.GetRequiredService<ILogger<Program>>();
-                    logger.LogError(e, "An error occured during migraiton");
+                await context.Database.MigrateAsync();
+
+                var env = services.GetRequiredService<IHostEnvironment>();
+                if (env.IsDevelopment())
+                {
+                    await Seed.SeedData(context);
                 }
             }
-
-            await host.RunAsync();
+            catch (Exception e)
+            {
+                var logger = services.GetRequiredService<ILogger<Program>>();
+                logger.LogError(e, "An error occured during migraiton");
+            }
         }
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder => {
-                    webBuilder.UseStartup<Startup>();
-                });
+        await host.RunAsync();
     }
+
+    public static IHostBuilder CreateHostBuilder(string[] args) =>
+        Host.CreateDefaultBuilder(args)
+            .ConfigureWebHostDefaults(webBuilder =>
+            {
+                webBuilder.UseStartup<Startup>();
+            });
 }
