@@ -4,38 +4,44 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authentication;
 using AspNet.Security.OAuth.Discord;
 using Microsoft.AspNetCore.Authentication.Cookies;
-using DisJockey.Extensions;
 using DisJockey.Shared.Extensions;
 
-namespace DisJockey.Controllers {
-    public class AccountController : BaseApiController {
+namespace DisJockey.Controllers
+{
+    public class AccountController : BaseApiController
+    {
 
         [HttpGet("login")]
-        public ActionResult Login() {
+        public ActionResult Login()
+        {
             var redirectUri = "/tracks";
 
             // if (Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT") == "Development") redirectUri = "https://localhost:4200" + redirectUri;
 
-            var challenge = Challenge(new AuthenticationProperties { RedirectUri = redirectUri }, DiscordAuthenticationDefaults.AuthenticationScheme );
+            var challenge = Challenge(new AuthenticationProperties { RedirectUri = redirectUri }, DiscordAuthenticationDefaults.AuthenticationScheme);
             return challenge;
         }
 
         [HttpGet("logout")]
-        public async Task<ActionResult> Logout() {
+        public async Task<ActionResult> Logout()
+        {
             await HttpContext.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
 
             return Redirect("/");
         }
-        
+
         [HttpGet("claims")]
-        public ActionResult<UserDto> GetUserInfo() {
+        public ActionResult<UserDto> GetUserInfo()
+        {
             var discordId = User.GetDiscordId();
 
-            if (discordId.HasValue) {
+            if (discordId.HasValue)
+            {
                 var username = User.GetUsername();
                 var avatarUrl = User.GetAvatarUrl();
 
-                var userDto = new UserDto {
+                var userDto = new UserDto
+                {
                     AvatarUrl = avatarUrl,
                     DiscordId = discordId.Value.ToString(),
                     Username = username
