@@ -1,10 +1,13 @@
 using System.IO;
 using Microsoft.AspNetCore.Mvc;
 
-namespace DisJockey.Controllers {
-    public class FallbackController : Controller {
-        public ActionResult Index() {
-            return PhysicalFile(Path.Combine(Directory.GetCurrentDirectory(), 
+namespace DisJockey.Controllers
+{
+    public class FallbackController : Controller
+    {
+        public ActionResult Index()
+        {
+            return PhysicalFile(Path.Combine(Directory.GetCurrentDirectory(),
                 "wwwroot", "index.html"), "text/HTML");
         }
     }
