@@ -1,20 +1,28 @@
 using DisJockey.Extensions;
 using DisJockey.Middleware;
+using DisJockey.Services;
+using DisJockey.Services.Interfaces;
+using MassTransit;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
+using System.Reflection;
 
-namespace DisJockey {
-    public class Startup {
+namespace DisJockey
+{
+    public class Startup
+    {
         private readonly IConfiguration _config;
-        public Startup(IConfiguration config) {
+
+        public Startup(IConfiguration config)
+        {
             _config = config;
         }
 
         // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services) {
+        public void ConfigureServices(IServiceCollection services)
+        {
             // Add Main WebApp Services (DB, Unit of Work)
             services.AddApplicationServices(_config);
 
@@ -27,11 +35,26 @@ namespace DisJockey {
 
             services.AddIdentityServices(_config);
 
-            services.AddDiscordServices(_config);
+            services.AddMassTransit(x =>
+            {
+                x.AddConsumers(Assembly.GetExecutingAssembly());
+
+                x.UsingSqlServer((context, cfg) =>
+                {
+                    cfg.ConfigureEndpoints(context);
+                });
+            });
+
+            services.Configure<SqlTransportOptions>(_config.GetSection("MassTransitSettings:SqlSettings"));
+
+            services.AddScoped<IDiscordTrackService, DiscordTrackService>();
+
+            //services.AddDiscordServices(_config);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env) {
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        {
             app.UseHttpsRedirection();
 
             app.UseRouting();
@@ -46,7 +69,8 @@ namespace DisJockey {
 
             app.UseMiddleware<ExceptionMiddleware>();
 
-            app.UseEndpoints(endpoints => {
+            app.UseEndpoints(endpoints =>
+            {
                 endpoints.MapControllers();
                 endpoints.MapFallbackToController("Index", "Fallback");
             });
